Modify dashboard mapping to take object values (#459)
Previously, logic related to dashboards would reference
TENSORBOARD_PLUGINS, a mapping between plugin name to dashboard element
name. Logic would also use the name of the plugin (the key of that
mapping) as the tab name displayed in TensorBoard.

This is suboptimal for several reasons - the plugin name is not
necessarily optimized for display purposes. For instance, take the
plugin name pr_curves. We would ideally remove the underscore from it.

We thus introduce the concept of DashboardData, which encapsulates data on a dashboard - the name of the element for the dashboard component (elementName), the name of the associated plugin (plugin), and the name to show within the navigation menu (tabName).

Renamed the _dashboards array within the tf-tensorboard component to
_dashboardInformationObjects to emphasize how that array is not a string
array, but rather contains objects.