Use `displayName` and `description` for PR curves
Summary:
Fixes #446.

The UI is the same as the existing UI as seen in the scalar dashboard
(including the existing display bug discussed in #430), but here’s a
screenshot, anyway:
![Screenshot.](https://user-images.githubusercontent.com/4317806/29992053-c10b61f8-8f60-11e7-95f2-36eef6058dd9.png)

Test Plan:
Load the PR curve dashboard, and ensure that you’re able to replicate
the screenshot above.

wchargin-branch: pr-curve-display-name-description