Add property nodeContextMenuItems to tf-graph
Added a property nodeContextMenuItems to the tf-graph and tf-graph-scene components. This optional property is an array of `contextmenu.ContextMenuItem`s. This property offers an API for developers to extend the context menu by adding custom items, each with a title and a callback.

See `getContextMenu` for an example array:

```ts
  let menu = [{
    title: (d): string => {
      return getIncludeNodeButtonString(node.include);
    },
    action: (elm, d, i) => {
      sceneElement.fire('node-toggle-extract', {name: node.name});
    }
  }];
```