Introduce tf-global-steps-selector
The PR curves plugin had used a special tf-pr-curve-steps-selector
component for letting the user select a single step to apply across all
tags for a given run.

This change moves that tf-pr-curve-steps-selector component into the
components directory and renames it to tf-global-steps-selector. Some
documentation was added to several properties of the component.

Also removed the tf-color-scale import from tf-pr-curve-dashboard. The
dashboard does not need it.