Log exceptions during processing in the modal
Previously, when an error occurred while the projector was processing data (creating data sets and vectors, normalizing data points, computing PCA, etc), it would be logged to the console.

This change uses a try ... catch block to make those exceptions appear as errors in the modal. This unfortunately hurts performance a bit because Chrome's V8 engine bypasses an optimized path for JS within a try ... catch block. However, so many people have taken issue with the lack of clarity surrounding projector bugs that I deem this change worth it.