Use flex positioning for card headings
In comments within #430, @wchargin proposed that we use flex positioning
to position elements within the header. That seems like a robust idea.

This PR thus uses @wchargin's proposed method to implement @jart's
original, space-efficient design of that makes use of both the left and
right sides of headings (if space is available).