Remove max op from numerator of precision (#465)
When the threshold is high enough so that the predictor makes no
positive predictions, precision is undefined because its denominator is
fp + tp. Hence, we remove the max operation from the numerator of the
precision calculation, so that precision is 0 (and not 1) when the thresold is
1.

A subsequent effort will make the frontend just not show data points for
which fp + tp = 0. For now, we get rid of that max op so that the user
does not see precision as equal to 1e-07, which can be confusing.

Also computed recall simply as

```python
recall = tp / tf.maximum(_MINIMUM_COUNT, tp + fn)
```

We already compute false positive counts via fn = tp[0] - tp. That seems
more straightforward.