Use flex positioning for card headings (#479)
In comments within #430, @wchargin proposed that we use flex positioning
to position elements within the header. That seems like a robust idea.

This PR thus uses @wchargin's proposed method to implement @jart's
original, space-efficient design of that makes use of both the left and
right sides of headings (if space is available).

This PR fixes the layout issues brought up in #430. This change also fixes
#421.