Remove the `demoify` function
Summary:
We don’t support the demo mode at all, nor do we have any plans to. This
function is complicating our logic and getting in the way of things that
we actually want to do, so out it goes.

Note that this commit does not remove “demo mode” entirely: the router
still has `demoMode`, and I didn’t go through and remove all of its
uses. (But that’s not a bad idea, and there are only a handful.)

Test Plan:
I tested that the image and audio dashboards still properly load images
and audio. Changes to the test are not tested, as we don’t run tests.

wchargin-branch: remove-demoify