Use compat to pre-empt how content is now bytes (#507)
A proto change will soon make the type of the content field of
PluginData 'bytes' instead of 'string'. This change pre-empts that
change by using the compat module to convert to bytes or text as
necessary within the debugger plugin.

With this commit, all tests pass under Python 2 and Python 3.5 on
TensorFlow nightly 2017-09-08.