Introduce the selected_dataset GET parameter (#505)
Introduced a selected_dataset GET parameter for the graph dashboard. The
value of this parameter is the name of the dataset (run) to load. This
parameter lets users send links to other users of the graph dashboard
that load specific datasets. Fixes #306.

When the user changes the dataset in the graph dashboard UI, the URL
parameter correspondingly changes. I tested this behavior out using the
various runs of the scalars demo. The correct dataset loads given the
selected_dataset parameter, and the URL changes when the dataset is
changed in the UI.

For instance, http://localhost:6006/#graphs&selected_dataset=temperature%3At0%3D270%2CtA%3D270%2CkH%3D0.001 makes the dashboard load the correct dataset.

If no dataset with the specified name can be found, we show a dialog
that notifies the user. This avoids misleading the user.

Set the type of the graph dashboard's `_datasets` property to Array. It had
previously been Object (incorrect).