Use `addParams` where it would improve code (#491)
Summary:
This commit replaces uses like `base + "?things"` (when `base` is not
known to not already contain a query component), or that fail to call
`encodeURIComponent` on the `things`, with uses of `addParams`.

I generated this commit by running `git grep '[&?].*='` and
investigating each occurrence. I ignored all occurrences in the graph
and projector plugins, as they have their own mini-ecosystems. I also
ignored the unique occurrence in the profile plugin, because (a) I can’t
test it, and (b) its inputs are hard-coded such that this change
provably would not fix any bugs.

Test Plan:
I verified the behavior of all existing plugins.

wchargin-branch: use-add-params