Stylize the newly selected node after expansion
Previously, when the user performs some functionality to expand the
graph and then select a node (such as navigating to a node after
observing TPU compatibility or clicking an op in the debugger plugin),
the graph would select a node before expanding its parent nodes.

This can cause an error because the selected node is not yet present in
the graph hierarchy:

```
(index):7546 Uncaught (in promise) TypeError: Cannot read property
'node' of undefined
   at HTMLElement._updateNodeState ((index):7546)
         at HTMLElement._selectedNodeChanged ((index):7547)
	 ```

This change makes the scene only stylize the selected node
after expansion of necessary nodes finishes. I don't think I can repro
the issue after this change.