Show TPU utilization% in a new column
Show TPU utilization% in a new column.
That column also shows the utilization color, and the time bar is now uniformly blue. This should reduce confusion about the two metrics that bar used to display.

Switched the rows to flexbox in order to get the layout right. (Unfortunately it can't be an actual table or display:table, due to the HTML nesting structure induced by polymer)