Only show sliders for runs with steps
The PR curves dashboard used to show a slider for each selected run.
A dysfunctional slider is still shown even if a run lacks PR curve data.

This change makes it so that only sliders for selected runs with PR
curve data are rendered.

Test plan: Load the PR curves dashboard. Toggle some runs. Try 10 times
to check for flakiness.