Load new projector config when dataset changes (#519)
Previously, when the data provider changed, a new projector config might not load because the name of the new run (string) could stay the same - indeed the name of the run is always 'proto'.

This change makes it so that we load a new projector config and correspondingly update the UI when the data provider changes (even if the run name changes).

Fixes #478. This commit is coauthored by @svsgoogle and @chihuahua.