Fix X and Y ranges of PR curves to [-0.05, 1.05]
Beforehand, the `vz-line-chart` component would set reasonable X and Y
ranges for all plots. This was problematic because as the user slides
across steps in the PR curves dashboard, the Y axis would change,
leading to difficulty in comparing across steps.

This change adds `defaultXRange` and `defaultYRange` properties to
`vz-line-chart` so that dashboards can manually set default ranges for
the X and Y axes if the need arises. It also makes `vz-line-chart` fire
a chart-rendered event after it renders - we can only reset the domain
of charts after that event is fired.

Test plan:
Run the PR curves demo, and start TensorBoard. Note that all PR curves
have X and Y ranges as specified by the commit summary.