Allow configuring pagination limit in settings (#535)
Summary:
In accordance with #483, this patch adds a configuration option to allow
changing the number of visualizations per page to a value other than 12.

Closes #483.

Test Plan:
Load the histogram dashboard with the histogram demo, which has seven
visualizations. Open the settings menu, and note that the default
pagination limit value is 12. Change the value to 1, close the dialog,
and move to page 6. Then, reopen the dialog and change the pagination
limit to 2; note that you are now on page 3 of 4. (This tests that the
existing-but-never-before-used behavior of `tf-pagination-view` that
attempts to retain approximately the same view across pagination limit
changes still works.)

Try to change the value to 0 or a negative value; note that this fails
without any console errors.

wchargin-branch: configurable-pagination-limit