Scan output base edges of functions
Previously, `cloneFunctionLibraryMetanode` would scan the outgoing edges of function outputs for destination nodes. This was problematic because much logic assumes that destination nodes are OpNodes - if the destination node happened to be a MetaNode instead, an error occured because no OpNode with a certain name could be found:

    :formatted:63999 Uncaught TypeError: Cannot read property 'inputs' of undefined

This change fixes that problem by instead scanning the base edges, which always have destinations that are OpNodes.

Test Plan: