Render subhierarchies of parents of functions
This change makes it so that when a node within the current subhierarchy being rendered has an input that is within a function call, we build the subhierarchy for the parent of that function call so that edges are computed correctly.

Edges for each hierarchy are rendered when the hierarchy is constructed. Hence, if we introduce the edges when the subhierarchies (containing the actual functions) render, it is too late for the higher-level hierarchy to render the edge.