Convert PR curve cards to use generalized charts (#568)
Refactored the card that loads PR curves to use the
tf-line-chart-loader component. This has several major advantages,
as detailed by #163, #189, and #204.

Part of this effort involved adding defaultXRange and
defaultYRange fields to tf-line-chart-data-loader.

Test plan:
Run the scalars and PR curves demos. Merge the directories, and
start TensorBoard. Note that the scalars and PR curves dashboards
still WAI. Check that sliders and buttons work.