Ensure that the list of function usages updates
Previously, when a user selected a node that represents a function, the list of function usages in the info card might not update. In response, this change

1. makes the info card asynchronously update that iron-list when the list of function usages changes. The logic for asynchronously updating node list items is now moved into a helper method also used by the logic for rendering lists of inputs and outputs.
2. makes `_functionUsages` depend on `_node` instead of `nodeName` like the other lists of nodes do. This prevents duplicating a little code and also ensure that `_resetState` is called when `_functionUsages` updates.
3. Renamed the 3 methods that obtain lists of nodes (`_getSuccessors`, `_getPredecessors`, and `_getFunctionUsages`) to be consistent. Moved them closer together so that future developers will hopefully see a pattern between them and not forget to asynchronously update the UI.

Test plan:
Start TensorBoard with the attached events file. In the graph dashboard, select an op node such as `y_`. Then select a function call node such as `MnistLoss` in the core graph. Notice that usages of the function are listed in the info card. Notice that the info card still lists inputs and successors for various ops throughout the graph as expected.