Remove add/remove button for library function nodes
Library functions are not part of the TensorFlow graph - they serve as templates for creating other ops. The graph explorer should thus not allow library functions to be extracted.

This change hides the 'Add/remove to/from the main graph' button from the info card when a library function is selected.