Make graph explorer work for functions that output to metanodes (#565)
Previously, if the output for a function was the input into an op within a metanode, the graph explorer would fail to visualize the graph and produce an error upon rendering the hierarchy for the metanode:

Uncaught TypeError: Cannot read property 'inputs' of undefined

This PR takes 2 actions to fix that.

1. We scan the base edges of the destination op, which have to have OpNodes as destinations (instead of scanning the inputs of the destination op, which could be MetaNodes).
2. If a node at the current hierarchy has an input that is within a function call, we render the hierarchy of the function call's parent so that edges are generated correctly (because the source nodes are now available).

Test plan:
Load the attached pbtxt file (See #565) into the graph explorer. Beforehand, the error would appear when the user expands the scope_foo metanode (because rendering an output op would fail). Now, the metanode correctly renders.