Add handleNodeSelected property to tf-graph
This change adds a `handleNodeSelected` property to the `tf-graph` and
`tf-graph-board` components. This property is a function that is called
every time the selected node changes (a new node is selected or the
current node is unselected).

This change is part of an effort to make the graph explorer
generalizable to various use cases.

Test plan: Modify the graph dashboard to pass a `handleNodeSelected`
callback to the `tf-graph-board` that console.logs selected nodes. After
starting TensorBoard, note that both node selection and deselection are
logged correctly.