Prevent library functions from being in/out extracted (#575)
Previously, the graph explorer would iterate through all nodes and deem nodes that lack any inputs or outputs to be in or out extracts. Library functions are isolated nodes within a special scene group, so they got lumped into the in and out extracts. This change adds a guard to prevent that.

Test Plan: Start TensorBoard with the attached events file. Add the init op to the main graph. Note that the library functions are no longer replicated, which had been a bug.