Add a streaming_op() to the pr_curves plugin. (#587)
This op produces the same result (a Tensor Summary containing the PR
curve data) as the existing op() function, but it does so by
aggregating the result over multiple batches of data.  The function
also returns an update_op that should be used to update the summary
for each batch of data.

With these changes, the function provides a tf.metrics-like function
for generating PR curve data.