Let user select edges if handleEdgeSelected set (#603)
Added a new property to `tf-graph` and `tf-graph-board` called
`handleEdgeSelected`. If this property is set, the user can select edges
by clicking on them. That same property is a callback run when an edge
is selected, allowing developers to handle edge selection.

Stylized selected edges so that they are red.

This change is part of an effort to generalize the graph explorer.

Test Plan: Pass a handleSelectedEdge property from `tf-graph-dashboard`
to `tf-graph-board`. Load a graph. Note that edges are now
selectable. Clicking on the graph deselects edges.

Also, try removing the handleEdgeSelected and restarting TensorBoard.
Note that edges are now not selectable.