Use registration pattern for dashboard tabs (#611)
The plugin list no longer needs to be specified at all. The HTML imports
themselves are now the plugin list. Special properties, e.g. custom tab names,
are now defined as properties on the Polymer components themselves.

The tradeoff is that each plugin's dashboard must call a registration function,
providing the plugin name and the name of the custom element.

See also: #459