Clarify DEVELOPMENT.md from a new dev perspective
This clarifies some things that I encountered - in particular, it specifies that you'll probably want TensorFlow installed in a virtualenv, and it updates the advice about getting the nightly build now that tf-nightly is on pypi (https://pypi.python.org/pypi/tf-nightly).  It also links to Bazel installation instructions just for the sake of being explicit.