Try just removing prefix if no match (#623)
Previously, the graph explorer attempted to extract a label from a
string to use as the display name for a node representing a function.
Apparently, that regex does not always match a function. Certain graphs
found within Google directly use the entire string as the name of the
function. That caused the graph explorer to be blank due to an error.

Now, the graph attempts to strip the prefix from the string and just
use the second half as the display name if parsing the display name
fails. This should be the less common case.

Test Plan: Start TensorBoard pointed at the events file in the PR. The
graph loads correctly as before. Also load the newly discovered internal
pbtxt. It also loads successfully.