Remove properties from tf-scalar-card (#627)
PR #626 had added default properties to vz-line-chart, and so we no
longer have to specify them in tf-scalar-card.

Specifically, we remove yValueAccessor and tooltipColumns but keep
the binding for xComponentsCreationMethod. We keep the latter to let
the user change the value on the horizontal axis.

Test Plan: Note that TensorBoard WAI on scalars demo data. Changing
horizontal axis (to wall time or relative) still works.