Add pb method for PR curves (#633)
This change adds a `pb` implementation for the PR curves summary,
which, like all `pb` implementations, lets users generate summaries
without having to use TensorFlow.

Also modified the test for PR curve summaries to use small test cases
that are easy for a developer to reason through instead of using the
demo data. This allows us to use the `compute_and_check_summary_pb`
paradigm for the PR curves summary test, just like for other plugins.

Updated the summary module to include `pr_curve_pb`. Fixes #445.

As part of this change, renamed the `tag` parameter of summary ops to
`name` to be consistent with other summaries.