Set _DEFAULT_NUM_THRESHOLDS to 201 (#645)
Set the default number of thresholds used by PR curve summary calculations to 201 instead of 200. 200 yielded strange, unfriendly bin edges. The number of bins is 1 less than the number of thresholds. Hence, it usually makes more sense for the number of thresholds to be odd.

This is a request from Waymo folks.