Log time taken by is_active() calls for plugin listing (#662)
This is a preliminary step in addressing tensorflow/tensorboard#625 that just adds logging (at INFO) for the amount of time taken by the is_active() call for each plugin loaded by TensorBoard when the plugins listing endpoint is invoked. This way as we make fixes we can monitor for regressions, and it also can help isolate the problematic plugin if we get reports from users that TensorBoard is taking a long time to respond with the list of active plugins (which is a prerequisite for basically any other usage of TensorBoard).

I also bundled in a couple small tweaks to other logging lines that can help diagnose slow plugin behavior.