Fix pageview hit counting to avoid dupes (#674)
This fixes some duplication issues in pageview counting for Google Analytics. Note that currently analytics is only enabled for internal Google users.

Specifically, sending pageview hits from _updateSelectedDashboard() was not really correct because that Polymer observer is invoked on any change to either observed property - the selected dashboard or the list of active dashboards. That means that as the active dashboards are loaded, every change to _activeDashboards will result in a duplicate page hit being logged for the currently selected dashboard.

As a special case of this, observers are also invoked a single time once their default values are all initialized to handle the default values (as a new change from "undefined"). So this also meant that on every page load we would send a pageview hit right away for the "/" page corresponding to the "null" default value of _selectedDashboard, even if loading e.g. http://tb/#scalars directly.

This fixes the first issue by changing _updateSelectedDashboard() to exclusively focus on setting the selected dashboard to the first active dashboard (if appropriate), and moving the pageview and hash updating into a separate observer _selectedDashboardChanged() that only observes _selectedDashboard. It fixes the second issue by making _selectedDashboard default to the value from the hash, so that loading http://tb/#scalars generates only a single pageview hit for "/scalars" rather than "/" followed by "/scalars".

Note that the way I have it now, loading http://tb/ will still always generate that first hit for "/" even if the active dashboards result in immediately "redirecting" to e.g. http://tb/#scalars which then gets recorded as a second hit. I think this is reasonable since the user did still type in "/" and that way we can distinguish visitors to the TensorBoard "home" page who get redirected to the first active dashboard from visitors who go immediately to a preselected dashboard.

Lastly, this removes the extra pageview hit for "/v/graph" sent by tf-graph-dashboard; it was introduced before the automatic hash-based plugin pageview hits were added to tf-tensorboard, and it's pretty much just redundant with the hit for "/graphs" that the latter would generate.