Allow for data series that are not runs
Some plugins (such as the `custom-scalar` and `debugger` plugins) make use of the `vz_line_chart` component but plot data series that are not runs (such as custom run-tag combos or tensor values).

This change makes the `tf-line-chart-data-loader` component accept a data series property that defaults to the list of runs.