Upgrade protobuf version in workspace.bzl (#655)
The previous version was very old (3.0.0), which may caused the
following error to occur when the --debugger_data_server_grpc_port
flag is used when launching the tensorboard binary from bazel run.
The version was also inconsistent with the protobuf dependency
version specified in pip_package/setup.py.

This change fixes those inconsistencies.
It also makes the protobuf version consistent with tensorboard.

Also in this change:
* Update grpcio version used by Travis CI.