Projector: Perturb t-SNE solution
A user-defined perturb factor determines the extent of random position
shifts upon pressing of Perturb button, which is useful to move out of
a local optimum or to alter the solution for a fresh perspective.
Perturb T-SNE by independently scaling point positions when Perturb
button is clicked, which is useful to move out of a local optima.