Projector: Semi-supervised t-SNE
Add to the projections-panel a supervise factor slider, an unlabeled
class specifier and a supervise column specifier. Capture the events
and update the dataset t-SNE variables that will be used to alter the
projections. Add a supervision clause to t-SNE to incorporate pairwise
prior probabilities based on label differences and similarities.