Set dataSeries to default to empty list
Make the default value for the dataSeries property an empty list.
Previously, PR #679 had set no default value for the property, which
meant that polymer would not trigger observers that depend on the
property, which meant that changing runs via the run selector did
nothing for dashboards that had not set the dataSeries property to a
non-falsy value (such as the scalars and PR curves dashboards).