Projector: Metadata editor
Add a metadata editor to the Projector, which gives the option to
modify attributes of selected points. Projector components related to
metadata display are refreshed after attribute changes, which also
expands the color palette for the modified attribute when a new class
is added. The main 'Label by' dropdown-menu is incorporated into the
metadata editor, which requires the user to view the point labels for
the metadata column being changed.