Include Roboto font in pip package (#751)
This change causes the Roboto fonts to be downloaded by Bazel and increases the
pip package size by 644kB. It uses the woff2 format, which is supported by all
major desktop browsers, except IE, and except Safari on old versions of Mac OS.

Please note that this only applies to western and western influenced character
sets. It's not possible to put all hundred megs of Noto CJK in the pip package.

This allows TensorBoard to look the same both offline and in countries whose
firewalls block Google Fonts.