Document the custom scalar plugin
We add a README.md file that documents how to collect data for the
plugin (log scalar data and set a layout) as well as how to use the
UI features of the dashboard.

We create a new images directory within the plugin that stores images
used in the documentation.