Projector: Selection edit with right-click
Now a mouse right click acts as selection editor, whether in normal or
editMode, such at it deselects a currently selected point (unless it is
the main/first point) and vice versa. Selection mode types are
introduced to support future expansion of different selection
interactions.

Add 'contextmenu' event capture in scatterplot, the standard HTML5
right-click event, and bind to selection 'edit' mode. This allows for
quick selection edits without explicitly switching to editMode via the
toolbar. This event handler checks that the event is not part of a drag
sequence or area select.