Projector: Selection edit with right-click (simplified)
Since the contextmenu event only fires after mousedown, we should
rather capture the right-click event during mouseup by checking for the
right mouse button. During mouseup, if it is not a drag sequence and
the right mouse button was clicked, we toggle selection/deselection of
the nearest point. This fixes the bug where right-click edited the
selection despite mousemove even though it was in fact part of a drag
sequence, as the logic failed earlier when contextmenu followed
mousedown as opposed to onclick which proceeds mousedown.