Projector: Custom projection (coordinate naming consistency)
Custom projection coordinates were named 'linear-x' and 'linear-y',
this was changed to 'custom-0' and 'custom-1' to allow for naming
consistency with pca, tsne and future projections.

This consistent format is then utilized to automate the extraction of
available projections from the datapoint coordinates, and auto-set the
distance metric/space choice correctly upon projection availability
changes.