Add TPU analysis reporting (#723)
This change allows the profile plugin to report the output of an internal TPU
input pipeline analysis tool. The report is only activated when a JSON file is
present under LOGDIR/plugins/profile/RUN/input_pipeline.json. The structure of
that file is described in tensorboard/plugins/profile/README.md.