Preload and precompress static assets in CorePlugin (#708)
This changes CorePlugin to load its static assets into memory at startup and precompress them with gzip. The two static assets in this case are "index.html" (the main vulcanized TensorBoard HTML blob) and "trace_viewer_index.html". Since "index.html" at a minimum is always served when using TensorBoard, this spends roughly 1 MB of extra RAM to move the cost of preparing it to startup-time so that it can be served directly from memory.

Prior to this change, a typical request for those assets would involve 1) a disk read to open the assets_zip_provider webfiles.zip file, 2) unzipping that zip file, 3) extracting the index.html file from the zip archive, and 4) recompressing the file using gzip (about 40ms to compress 2.6MB down to 0.7ish) to send it back to the client.

Now it's just served as precompressed gzip data directly from memory. From some basic testing on my workstation with a local TensorBoard, this changes the TTFB (Time To First Byte [1]) for index.html from about 100ms average to less than 5ms, and the average delay to DOMContentLoaded firing correspondingly drops from about 1000ms to 900ms.

Modern browsers pretty much all send Accept-Encoding: gzip but in case the user doesn't send that header, this change also adds support for decompressing precompressed data before responding. In theory we could avoid the decompression delay as well by also storing uncompressed assets in memory, but I opted not to because it complicates the Respond() API and the argument in favor is much weaker since decompression is typically faster than compression (in testing it took about 20ms for index.html) and we expect gzip supporting requests to be the vast majority, so it'd be spending more memory to save less time less often.

[1] https://developers.google.com/web/tools/chrome-devtools/network-performance/reference#timing-explanation