Projector: Inspector metadata context
Add a metadata context to the inspector panel, showing the available
labels and label counts for the 'Edit by' metadata column. By clicking
a label in this list it is immediately applied to the point selection,
which makes it easier to modify metadata with longer labels since they
do not have to be entered every time.

Metadata context is entered whenever there is content in the metadata
editor label field, as soon as the field is cleared the context is
removed. While in metadata context the neighborhood size can still be
changed, and a label search can be performed to highlight points with a
certain label not necessarily from the same metadata column. This adds
versatility in locating point groups with a specified label while being
able to quickly modify the metadata.