Make debugger plugin backdrop not block navbar (#806)
We make the debugger plugin use a custom backdrop that does not occlude
the main TensorBoard navigation bar. This lets the user navigate to a
different dashboard after arriving at the debugger dashboard and
encountering the dialog that notes how the plugin is waiting for the
first session run.

Fixes #793.