Correctly render function nodes
This PR fixes some issues that enable users to examine the intricacies
of nodes that represent functions:

* We only patch outputs of functions at the very end of cloning a
metanode that encapsulates its internal content. Previously, the graph
explorer patched edges from output nodes that did not exist.
* We render subhierarchies of metanodes only after extracting nodes from
the graph. Otherwise, the graph logic errs because it later tries to
construct edges to nodes that would later be extracted from the graph.
* We take into account how the last few inputs into a function might map
to the same node. Otherwise, edge-patching logic might reference a
non-existent node.